// src/lib/types.ts
export interface Sprocket {
  id: number
  chain_size: string
  standard: string
  chain_pitch: string
  chain_pitch_inches: number
  num_teeth: number
  outside_diameter: string
  outside_diameter_inches: number
  bore_type: string
  bore_size_range: string
  mcmaster_part_number: string
  price: number
}